requirements.txt
Flask==2.0.1
Flask-Login==0.5.0
flask-wtf==1.1.1
WTForms==3.0.1
Werkzeug==2.0.1
python-dotenv==0.19.0
email-validator==1.1.3
APScheduler==3.8.1
qrcode==7.3
Pillow==9.5.0
reportlab==3.6.8
pandas==1.3.3
openpyxl==3.0.9
xlsxwriter==3.0.2
SQLite3-api==2.0.1